% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.ITI.R
\name{drop.ITI}
\alias{drop.ITI}
\title{Drop ITI trials from an alex data.table}
\usage{
drop.ITI(d)
}
\arguments{
\item{d}{An alex data.table}
}
\value{
An alex data.table with ITI rows removed
}
\description{
Drop ITI trials from an alex data.table
}

