% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.ITI.responders.R
\name{drop.ITI.responders}
\alias{drop.ITI.responders}
\title{Remove participants responding too often during inter-trial
intervals (ITIs), whih potentially signals lack of attention
to the experiment).}
\usage{
drop.ITI.responders(d, threshold, drop = TRUE)
}
\arguments{
\item{d}{An alex data.table}

\item{threshold}{Fraction of ITI trials above which a participant is flagged for removal}

\item{drop}{If TRUE, the flagged participants are removed, if FALSE they are not}
}
\value{
With drop=TRUE, return the data.table after removing all rows pertaining to the flagged participants. When drop=FALSE, return the original data.table
}
\description{
Remove participants responding too often during inter-trial
intervals (ITIs), whih potentially signals lack of attention
to the experiment).
}

