% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drop.phases.R
\name{drop.phases}
\alias{drop.phases}
\title{Remove data that come from one or more experimental phases}
\usage{
drop.phases(d, unwanted)
}
\arguments{
\item{d}{A data frame returned by \code{read.alex}}

\item{unwanted}{A list of names of experimental phases to be removed}
}
\value{
A data frame without data rows from the unwanted phases

A copy of the original data frame after dropping the relevant rows
}
\description{
Remove data that come from one or more experimental phases
}
\examples{
my.data <- drop.phases( d, c("Instructions") )
}

