% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.S2.responses.R
\name{drop.S2.responses}
\alias{drop.S2.responses}
\title{Drop responses that occurred during the S2}
\usage{
drop.S2.responses(d)
}
\arguments{
\item{d}{An alex data frame}
}
\value{
A copy of the original data frame after dropping the relevant rows
}
\description{
Drop responses that occurred during the S2
}
