% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.the.droppable.R
\name{drop.the.droppable}
\alias{drop.the.droppable}
\title{Combines \code{drop.phases}, \code{drop.S2.responses},
\code{drop.ITI}, and \code{drop.too.fast}}
\usage{
drop.the.droppable(d, phases = c(), rt = 100, iti = TRUE)
}
\arguments{
\item{d}{An alex data frame}

\item{phases}{A list of phases for \code{drop.phases}}

\item{rt}{Reaction time threshold (see \code{drop.too.fast})}

\item{iti}{Drop responses during ITI if \code{TRUE}}
}
\value{
A copy of the original data frame after dropping the relevant rows
}
\description{
Combines \code{drop.phases}, \code{drop.S2.responses},
\code{drop.ITI}, and \code{drop.too.fast}
}
