% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drop.too.fast.R
\name{drop.too.fast}
\alias{drop.too.fast}
\title{Drop responses with fast reaction times. All occurring before
\code{threshold} ms (default: 100 ms) following trial start are
dropped.}
\usage{
drop.too.fast(d, threshold = 100)
}
\arguments{
\item{d}{An alex data frame}

\item{threshold}{Response threshold}
}
\value{
A copy of the original data frame after dropping the relevant rows
}
\description{
Drop responses with fast reaction times. All occurring before
\code{threshold} ms (default: 100 ms) following trial start are
dropped.
}

