% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.alex.R
\name{is.alex}
\alias{is.alex}
\title{Check whether an object is an alex data.table}
\usage{
is.alex(d)
}
\arguments{
\item{d}{an R object}
}
\value{
TRUE/FALSE
}
\description{
Check whether an object is an alex data.table
}

