% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation.and.keynum.R
\name{presentation.and.keynum}
\alias{presentation.and.keynum}
\title{Build 'Presentation' and 'KeyNum' variables for alex data. The
first is the trial number for a given stimulus and phase, the
second numbers responses within presentations.}
\usage{
presentation.and.keynum(dt)
}
\arguments{
\item{data}{An alex data.table}
}
\value{
Nothing, the original data.table is modified in place to have KeyNum and Presentation variables
}
\description{
Build 'Presentation' and 'KeyNum' variables for alex data. The
first is the trial number for a given stimulus and phase, the
second numbers responses within presentations.
}
\examples{
This function is meant for internal use
}
