% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.alex.R
\name{read.alex}
\alias{read.alex}
\title{Read alex data residing in a Data directory under \code{data.dir}. Presentation
and KeyNum variables are added. The first is the number of the current presentation
for the current stimulus (as opposed to Trial, which spans across stimuli). The second is
the number of the current response in the current trial. See the alex manual (look for
AlexManual.pdf in your alex installation, or use the command alex-manual, or look online
at http://github.com/drghirlanda/alex). A GSubject (global subject) identified is also added
by merging group and subject identifiers.}
\usage{
read.alex(data.dir = getwd())
}
\arguments{
\item{data.dir}{The directory where the alex Data folder resides (default: the working directory)}
}
\value{
a data.table containing all read data
}
\description{
Read alex data residing in a Data directory under \code{data.dir}. Presentation
and KeyNum variables are added. The first is the number of the current presentation
for the current stimulus (as opposed to Trial, which spans across stimuli). The second is
the number of the current response in the current trial. See the alex manual (look for
AlexManual.pdf in your alex installation, or use the command alex-manual, or look online
at http://github.com/drghirlanda/alex). A GSubject (global subject) identified is also added
by merging group and subject identifiers.
}
\examples{
my.data <- read.alex()
my.data <- read.alex( "Some/Directory" )
}

