% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regroup.R
\name{regroup}
\alias{regroup}
\title{Rename experimantal groups in an alex data.table}
\usage{
regroup(d, gList)
}
\arguments{
\item{d}{An alex data.table}

\item{gList}{a list of the form list( new1=old1, new2=old2, ... ) where new1, new2, etc., are new group labels, and old1, old2, etc., are old group labels (scalars or vectors) to be renamed)}
}
\value{
Nothing, the data.table is modified in-place
}
\description{
Rename experimantal groups in an alex data.table
}
\examples{
regroup( d, list( Training=c("Training1", "Training2") ) )
}

