% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.rates.R
\name{response.rates}
\alias{response.rates}
\title{Calculate mean response rates per S1, Group, Presentation, and Phase}
\usage{
response.rates(d, by.subject = FALSE)
}
\arguments{
\item{d}{An alex data frame}
}
\value{
A data frame with Rate and RateSE columns added, the second
  being the standard error of the response rate
}
\description{
Calculate mean response rates per S1, Group, Presentation, and Phase
}
\examples{
d <- read.alex()
r <- mean.response.rates( d )
}

